Refactor usagestats functions
